sql assistant logic moved to a dedicated file
